Extracted common functionality to base class
